Server/requirements.txt
flask==2.0.3
flask-cors
flask-jwt-extended
pymongo
confluent-kafka
flask-mongoengine==1.0.0
python-dotenv
werkzeug==2.0.3